Fix bug with memcpy

+++ utils/config.h
@@ -47,7 +47,7 @@ typedef struct packet{
     uint16_t type;               //Tipo do pacote(p.ex. DATA| CMD)
     uint16_t seqn;               //Número de sequência
     uint32_t total_size;         //Número total de fragmentos
-    uint16_t length;             //Comprimento do payload 
+    uint32_t length;             //Comprimento do payload 
     char _payload[PAYLOAD_SIZE]; //Dados do pacote
     int payloadCommand;
 } packet;
